refactor: group all functions only used by config/utils in one file
