+++ everyvoice/utils/__init__.py
@@ -13,7 +13,6 @@ from unicodedata import normalize
 
 import yaml
 from loguru import logger
-from pydantic import ValidationInfo
 
 from everyvoice import exceptions
 from everyvoice.config.type_definitions import TargetTrainingTextRepresentationLevel
@@ -72,7 +71,10 @@ def filter_dataset_based_on_target_text_representation_level(
     match target_text_representation_level:
         case TargetTrainingTextRepresentationLevel.characters:
             target_training_text_key = "character_tokens"
-        case TargetTrainingTextRepresentationLevel.ipa_phones | TargetTrainingTextRepresentationLevel.phonological_features:
+        case (  # noqa: E211
+            TargetTrainingTextRepresentationLevel.ipa_phones
+            | TargetTrainingTextRepresentationLevel.phonological_features
+        ):
             target_training_text_key = "phone_tokens"
         case _:
             raise NotImplementedError(
@@ -170,87 +172,6 @@ def update_config_from_cli_args(arg_list: List[str], original_config):
             expand_config_string_syntax(arg)
         )
     return original_config
-
-
-def relative_to_absolute_path(
-    value: Any, info: Optional[ValidationInfo] = None
-) -> Path | None:
-    """
-    Helper function to annotate a type.
-    This function processes relative paths and either resolve them to absolute
-    paths or resolve them with respect to the configuration file they came
-    from.
-    """
-    if value is None:
-        return value
-
-    try:
-        # Make sure value is a path because it can be a string when we load a
-        # model that is not partial.
-        path = Path(value)
-        if (
-            not path.is_absolute()
-            and info
-            and info.context
-            and (config_path := info.context.get("config_path", None))
-        ):
-            path = (config_path.parent / path).resolve()
-        return path
-    except TypeError as e:
-        # Pydantic needs ValueErrors to raise its ValidationErrors
-        raise ValueError from e
-
-
-def directory_path_must_exist(
-    value: Any, info: Optional[ValidationInfo] = None
-) -> Path | None:
-    """
-    Helper function to annotate a type.
-    Creates a directory if it doesn't exist.
-    """
-    assert isinstance(value, Path)
-    if (
-        info
-        and info.context
-        and (writing_config := info.context.get("writing_config", None))
-    ):
-        # We are writing the original config and must temporarily resolve the path.
-        (writing_config.resolve() / value).mkdir(parents=True, exist_ok=True)
-    else:
-        if not value.exists():
-            logger.info(f"Directory at {value} does not exist. Creating...")
-            value.mkdir(parents=True, exist_ok=True)
-
-    return value
-
-
-def path_is_a_directory(
-    value: Any, info: Optional[ValidationInfo] = None
-) -> Path | None:
-    """
-    Helper function to annotate a type.
-    Verifies ala `PathType("dir")` that `value` is a directory.
-    """
-    if (
-        info
-        and info.context
-        and (writing_config := info.context.get("writing_config", None))
-    ):
-        # We are writing the original config and must temporarily resolve the path.
-        tmp_path = writing_config.resolve() / value
-        if not tmp_path.is_dir():
-            raise ValueError(f"{tmp_path} is not a directory")
-    else:
-        try:
-            # Make sure value is a path because it can be a string when we load a model that is not partial.
-            path = Path(value)
-            if not path.is_dir():
-                raise ValueError(f"{path} is not a directory")
-        except TypeError as e:
-            # Pydantic needs ValueErrors to raise its ValidationErrors
-            raise ValueError from e
-
-    return value
 
 
 def original_hifigan_leaky_relu(x):
@@ -407,6 +328,7 @@ def generic_dict_loader(
         list[dict]: a list of dicts representing the rows in the filelist
     """
     assert fieldnames is not None or file_has_header_line
+    f: Iterable[str]
     with open(path, "r", newline="", encoding="utf8") as f:
         if record_limit:
             f = islice(f, record_limit)
